Bug fixes and improvements
CEXAPI.cs:
Fixed an issue where the nonce was not correct, I used the day time,
which would throw errors when you try and run it earlier than the day
before. Changed to use total MS from start of time.
Added error return handling from the API call. A new ApiException object
will be thrown now when an error is received from the API call

ApiObjects\*
Updated some of the formatting for the ToString methods of some classes
Added an overloader to ToString for AccountBalanceTrade.cs. This will
now either include or exclude current orders in the resulting string